pyproject.toml
[project]
name = "reflex-chef"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "groq>=0.19.0",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "pyjwt>=2.10.1",
    "reflex>=0.7.2",
    "rich>=13.9.4",
]
